finished navbar,sidebar,changed css a little
