+++ src/App.js
@@ -11,6 +11,7 @@ import StudyInfoPage from "./pages/StudyInfoPage/StudyInfoPage";
 import StudyRecruitPage from "./pages/StudyRecruitPage/StudyRecruitPage";
 import MyInfoPage from "./pages/MyInfoPage/MyInfoPage";
 import AllStudiesPage from "./pages/AllStudiesPage/AllStudiesPage";
+import DifferentLayout from "./layout/DifferentLayout";
 
 
 
@@ -29,11 +30,13 @@ function App() {
             <Route index element={<MyStudyPage/>}/> {/*내 스터디 페이지*/}
             <Route path=":id" element={<StudyInfoPage/>}/> {/*스터디 상세 페이지*/}
           </Route>
-          <Route path ="newstudy" element={<StudyCreatePage/>}/> {/*스터디 생성 페이지*/}
+          
           <Route path="login" element={<LoginPage/>}/>
           <Route path="signup" element={<SignupPage/>}/>
           <Route path="myinfo" element={<MyInfoPage/>}/>
         </Route>
+        
+        <Route path ="/newstudy" element={<><DifferentLayout/><StudyCreatePage/></>}/> {/*스터디 생성 페이지*/}
 
         <Route path="*" element={<NotFoundPage />} />
       </Routes>
