--- src/default/SideBar.css
@@ -1,39 +1,0 @@
-.sidebar {
-  height: 100vh;
-  border-right: 0.1px solid #c7c7c7;
-  background-color: #fafafa;
-  padding: 0.8rem;
-  color: #4c4e33;
-}
-
-.sidebarList {
-  list-style: none;
-  padding: 0.5rem;
-}
-
-.sidebarMenu {
-  margin-bottom: 2rem;
-}
-
-.sidebarListItem {
-  padding: 1rem;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  border-radius: 0.3rem;
-  margin-bottom: 0.6em;
-}
-
-.sidebarListItem.active,
-.sidebarListItem:hover {
-  background-color: #e8e8e8;
-}
-
-.sidebarListItem:first-child {
-  border-bottom: 1px solid #c7c7c7;
-}
-
-.sidebarIcon {
-  margin-right: 0.6rem;
-  font-size: 20px !important;
-}
